pacat-simple-vchan: fix broken check for errors
Assigning an error code to a size_t ensures it will never be
seen as negative, but as very large. If an error happens on
the first call, the buffer pointer would be rewound one byte
(pointing outside the actual buffer).

If the buffer happens to have bytes available that span the
end of the circular buffer, but the notify call fails (which
I'm not convinced it can if the function is called), then
there will still be data available on the second loop run,
and this will write one byte before the buffer pulse gave us,
or read one byte before (depending on the function this
happens in).

In the case of send_rec_data, since each loop would increase
rec_buffer_length, we'd get an infinite (or close enough)
loop, wedging guid (till it gets killed by SIGSEGV).

+++ pulse/pacat-simple-vchan.c
@@ -155,7 +155,7 @@ static void start_drain(struct userdata *u, pa_stream *s) {
 
 static void process_playback_data(struct userdata *u, pa_stream *s, size_t max_length)
 {
-	size_t l = 0;
+	int l = 0;
 	size_t index, buffer_length;
 	void *buffer = NULL;
 	assert(s);
@@ -211,7 +211,7 @@ static void stream_write_callback(pa_stream *s, size_t length, void *userdata) {
 }
 
 static void send_rec_data(pa_stream *s, struct userdata *u) {
-	size_t l;
+	int l;
 	const void *rec_buffer;
 	size_t rec_buffer_length;
 	int rec_buffer_index;
